requirements.txt
Django==1.4
psycopg2==2.4.5
South==0.7.4
django-ratings==0.3.6
tagging==0.2.1
django-tastypie==0.9.11
django-taggit==0.9.3